+++ src/test/kotlin/ru/yoomoney/tech/grafana/dsl/metrics/prometheus/PrometheusMetricTest.kt
@@ -15,6 +15,17 @@ class PrometheusMetricTest {
     }
 
     @Test
+    fun `builder should create instantVector with regex-match`() {
+        val instantVector = instantVector(
+                metricName = "metricName",
+                exactlyLabels = linkedMapOf("label_1" to "value_1", "label_2" to "value_2"),
+                regexLabels = linkedMapOf("label_reg" to ".*")
+        )
+
+        instantVector.asString() shouldBeEqualTo """metricName{label_1="value_1", label_2="value_2", label_reg=~".*"}"""
+    }
+
+    @Test
     fun `builder should create rangeVector`() {
         val instantVector = rangeVector(
             metricName = "metricName",
@@ -23,5 +34,17 @@ class PrometheusMetricTest {
         )
 
         instantVector.asString() shouldBeEqualTo """metricName{label_1="value_1", label_2="value_2"}[1h]"""
+    }
+
+    @Test
+    fun `builder should create rangeVector with regex-match`() {
+        val instantVector = rangeVector(
+                metricName = "metricName",
+                exactlyLabels = linkedMapOf("label_1" to "value_1", "label_2" to "value_2"),
+                regexLabels = linkedMapOf("label_reg" to ".*"),
+                interval = "1h"
+        )
+
+        instantVector.asString() shouldBeEqualTo """metricName{label_1="value_1", label_2="value_2", label_reg=~".*"}[1h]"""
     }
 }